# Improve SliceExt::binary_search performance