@bluss 2017-10-17
I like it, trading an unpredictable Branch for a predictable one (bound check) is neat.

It doesn't exit when it finds the first equal element -- so does this PR change the result for some inputs?