@arthurprs 2017-10-17
Cool! Although I think this needs to be tested further with other payloads, like (u64, u64), where the comparison isn't compiled into a single `cmov*`.