@alkis 2017-10-17
@arthurprs I don't think anything will change if (u64, u64) is used (or even strings for that matter). The `cmov` is used to select the new `base` after the comparison is done and its result is placed in the flags register: it is either `base` or `mid`. If the comparison operator has branches of its own that are not predictable there is not much we can do about it at this level. It has to be done by code external to `binary_search_by`.