@arthurprs 2017-10-17
Good point. I also ran a few benches just to confirm and it checks out.