@bluss 2017-10-17
@alkis there are no test cases that cover this — with duplicate elements. I've had time to check this at my computer now, and the new algorithm does produce different results for some inputs:

The example [(playground link)](https://play.rust-lang.org/?gist=a97e252cdad440c6b75525ef6d31d55e&version=nightly) is searching for 0 in a vector of zeros.

1. It's important to walk into breaking changes, however small, with open eyes
2. Now we know that it exists — can we find out if it matters? We've had bugs filed for breakage for less.

(Less important, but still interesting, is that the old algorithm will finish as quickly as possible with this input and the new algorithm it's a worst case.)