@alkis 2017-10-17
@bluss: thanks for testing this. I don't think these changes are material and very unlikely to result in breakages. Let me explain why:
* old: returns the first match. The match is implementation defined. Also it is not a well defined element: it is neither the lower bound nor the one before the upper bound. It is hard to make a case for using such match with any reasonable expectations.
* new: if a match exists it returns the lower bound. This is a stronger guarantee than the old implementation. This is an implementation detail but it has the risk of falling into [Hyrum's Law](http://www.hyrumslaw.com/) in the future.

In addition the two implementations will have different behavior when the list is not sorted given the partial order defined by the predicate. I also think that this is not important.

For performance it is true that the new algorithm always does `log2(size)` steps but the old does at most `log2(size)` steps. This is both good and bad. Bad because it can be slower in contrived situations and good because it has predictable performance.

I can add a few more benchmarks if you think they add value:
* string slices as keys
* with dups
* with different sizes: sizeof array fits in L1 cache, L2 cache and L3 cache

About possible breakages, if you have suggestions on how to investigate them before merging I would be glad to take a look.