@alkis 2017-10-17
I improved the benchmarks a bit. Please take another look.