@arthurprs 2017-10-18
I found myself wanting the lower bound of the equal subsequence recently. On the other hand I'd argue against guaranteeing this sort of behavior. Also, this discussion resembles the stable/unstable sort thing.