@kennytm 2017-10-18
The [documentation of `binary_search`](https://doc.rust-lang.org/std/primitive.slice.html#method.binary_search) never guaranteed which index will be returned when there's an equal range. The new behavior is compatible with the existing definition.

> Looks up a series of four elements. The first is found, with a uniquely determined position; the second and third are not found; <ins>**the fourth could match any position in `[1, 4]`.**</ins>
>
> ```rust
> let s = [0, 1, 1, 1, 1, 2, 3, 5, 8, 13, 21, 34, 55];
> 
> assert_eq!(s.binary_search(&13),  Ok(9));
> assert_eq!(s.binary_search(&4),   Err(7));
> assert_eq!(s.binary_search(&100), Err(13));
> let r = s.binary_search(&1);
> assert!(match r { Ok(1...4) => true, _ => false, });
> ```