@alkis 2017-11-01
When merging please use the following message instead of the original.

---


Improve the performance of binary_search by reducing the number of unpredictable conditional branches in the loop.

Before:

```
test slice::binary_search_l1                               ... bench:          48 ns/iter (+/- 1)
test slice::binary_search_l2                               ... bench:          63 ns/iter (+/- 0)
test slice::binary_search_l3                               ... bench:         152 ns/iter (+/- 12)
test slice::binary_search_l1_with_dups                     ... bench:          36 ns/iter (+/- 0)
test slice::binary_search_l2_with_dups                     ... bench:          64 ns/iter (+/- 1)
test slice::binary_search_l3_with_dups                     ... bench:         153 ns/iter (+/- 6)
```

After:

```
test slice::binary_search_l1                               ... bench:          17 ns/iter (+/- 1)
test slice::binary_search_l2                               ... bench:          24 ns/iter (+/- 2)
test slice::binary_search_l3                               ... bench:         139 ns/iter (+/- 27)
test slice::binary_search_l1_with_dups                     ... bench:          17 ns/iter (+/- 1)
test slice::binary_search_l2_with_dups                     ... bench:          25 ns/iter (+/- 1)
test slice::binary_search_l3_with_dups                     ... bench:         137 ns/iter (+/- 21)
```